+++ bigmler/tests/project_creation_steps.py
@@ -24,7 +24,7 @@ import json
 from bigmler.tests.world import world, res_filename
 from subprocess import check_call, CalledProcessError
 from bigmler.checkpoint import file_number_of_lines
-from bigmler.utils import SYSTEM_ENCODING, PYTHON3
+from bigmler.utils import BIGML_SYS_ENCODING, PYTHON3
 from bigml.api import check_resource, BigML
 from bigmler.tests.common_steps import check_debug
 from nose.tools import assert_equal, assert_not_equal, ok_
@@ -48,7 +48,7 @@ def i_create_source_with_project(step, data=None, project=None, output_dir=None)
                    output_dir +
                    u" --project=\"" + project + "\"")
         if not PYTHON3:
-            command = command.encode(SYSTEM_ENCODING)
+            command = command.encode(BIGML_SYS_ENCODING)
         command = check_debug(command)
         retcode = check_call(command, shell=True)
         if retcode < 0:
@@ -69,7 +69,7 @@ def i_create_project_in_org(step, name=None, output_dir=None, organization=None)
                    "\" --organization " + organization +
                    u" --output-dir " + output_dir)
         if not PYTHON3:
-            command = command.encode(SYSTEM_ENCODING)
+            command = command.encode(BIGML_SYS_ENCODING)
         command = check_debug(command)
         retcode = check_call(command, shell=True)
         if retcode < 0:
@@ -91,7 +91,7 @@ def i_create_source_with_org_project(step, data=None, output_dir=None):
                    output_dir +
                    u" --org-project " + world.project["resource"])
         if not PYTHON3:
-            command = command.encode(SYSTEM_ENCODING)
+            command = command.encode(BIGML_SYS_ENCODING)
         command = check_debug(command)
         retcode = check_call(command, shell=True)
         if retcode < 0:
@@ -113,7 +113,7 @@ def i_create_source_with_project_id(step, data=None, output_dir=None):
                    output_dir +
                    u" --project-id " + world.project['resource'])
         if not PYTHON3:
-            command = command.encode(SYSTEM_ENCODING)
+            command = command.encode(BIGML_SYS_ENCODING)
         command = check_debug(command)
         retcode = check_call(command, shell=True)
         if retcode < 0:
@@ -162,7 +162,7 @@ def i_create_project(step, project=None, output_dir=None):
                    u"\" --store --output-dir " +
                    output_dir)
         if not PYTHON3:
-            command = command.encode(SYSTEM_ENCODING)
+            command = command.encode(BIGML_SYS_ENCODING)
         command = check_debug(command)
         retcode = check_call(command, shell=True)
         if retcode < 0:
@@ -185,7 +185,7 @@ def i_update_project(step, params=None, values=None):
             command += u" --%s %s" % (param, value)
 
         if not PYTHON3:
-            command = command.encode(SYSTEM_ENCODING)
+            command = command.encode(BIGML_SYS_ENCODING)
         command = check_debug(command)
         retcode = check_call(command, shell=True)
         assert retcode >= 0