+++ HISTORY.rst
@@ -3,9 +3,11 @@
 History
 -------
 
-3.26.4 (2020-04-27)
+3.27.0 (2020-05-26)
 ~~~~~~~~~~~~~~~~~~~
 
+- Adding bigmler connector subcommand.
+- Fixing issues with --resources-log and --clear-logs in Python 3.X.
 - Fixing encoding issues when using non-ascii characters in paths.
 
 3.26.3 (2020-04-17)